requirements.txt
pyaudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1
numpy==2.1.3
openai-whisper==20240930
pyannote.audio==3.3.2
anthropic==0.25.1
python-dotenv==1.0.1
torch==2.5.1
torchaudio==2.5.1
scipy==1.14.1